[v 1.0.0 ] [fix: 모듈 구조 피드백 반영 ]
