+++ iTime/Feature/Sources/AppRootImpl/AppRootRouter.swift
@@ -7,6 +7,7 @@
 
 import RIBs
 
+import AppRoot
 import LoggedOut
 import LoggedIn
