+++ iTime/Platform/Sources/NetworkRepositoryImpl/FirestoreRepository/Protocol/iTimeSnapShot.swift
